requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
comtypes==1.4.13
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
numpy==2.3.5
openai==2.8.1
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pypiwin32==223
python-dotenv==1.2.1
pyttsx3==2.99
pywin32==311
sniffio==1.3.1
sounddevice==0.5.3
SpeechRecognition==3.14.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
